Rename: popup, content 하위에 component 폴더 추가
