+++ extension/src/content/components/UserFilter.tsx
