+++ extension/src/content/components/index.ts
@@ -1,0 +1,1 @@
+export { UserFilter } from './UserFilter'
