+++ extension/src/popup/components/PopupErrorFallback.tsx
