+++ extension/src/popup/components/index.ts
@@ -1,0 +1,2 @@
+export { Popup } from './Popup'
+export { PopupErrorFallback } from './PopupErrorFallback'
